requirements.txt
amrlib==0.7.0
arxiv
arxiv2bib
boto3==1.9.118
bert-extractive-summarizer
joblib
keybert
numpy
pandas
pdftotext
Pillow==9.1.1
PyMuPDF
pdfminer3k
requests
scholarly
scikit_learn
sentence_transformers
sentencepiece
setuptools
spacy
scispacy
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.0/en_core_sci_scibert-0.5.0.tar.gz
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.0/en_core_sci_lg-0.5.0.tar.gz
streamlit
streamlit-pydantic
summarizer
tabula
tabula_py
tensorflow
tensorflow_hub
torch
transformers
unicodedata2
